requirements.txt
numpy==1.11.3
scipy==0.19.0
pandas==0.19.2
bokeh==0.11.1
seaborn==0.7.1
sklearn
Flask==0.10.1
Flask-OAuthlib==0.9.4
requests==2.2.1
bs4==0.0.1
readability-lxml==0.6.2
